Added tests for L=32 and fixed the problems discovered doing this testing and updated old tests

+++ src/main/java/compression/utility/BitUtil.java
@@ -17,7 +17,8 @@ public class BitUtil {
     }
 
     public static int bits2Int(String bits) {
-        return Integer.parseInt(bits, 2);
+        // Necessary hack to parse 1111 1111 1111 1111 1111 1111 1111 1111
+        return (int)(Long.parseLong(bits, 2));
     }
 
     /**
